Use ArcticRLClient after the main client rename.
The rebase onto #83 dropped SyncArcticRLClient / ArcticRLClientConfig; recipes now construct ArcticRLClient(ArcticClientConfig).
