+++ recipes/rl/trl/txt2sql/run_qwen3_1.7b_bird_grpo_arl.py
@@ -70,11 +70,11 @@ def build_client(model: str, host: str, port: int, *, max_seq_len: int, response
                  tensor_parallel_size: int, comm_protocol: str = "http", seed: int = 42,
                  colocate: bool = False, zorro_train_enable: bool = False, gpu_mem_util: float = 0.3,
                  grad_accum_steps: int = 1):
-    from arctic_platform.client.client import SyncArcticRLClient
-    from arctic_platform.client.config import ArcticRLClientConfig
-    from arctic_platform.client.config import OnPremConfig
-    from arctic_platform.client.config import SamplingConfig
-    from arctic_platform.client.config import TrainingConfig
+    from arctic_platform.client import ArcticClientConfig
+    from arctic_platform.client import ArcticRLClient
+    from arctic_platform.client import OnPremConfig
+    from arctic_platform.client import SamplingConfig
+    from arctic_platform.client import TrainingConfig
 
     ds_config = {
         "train_micro_batch_size_per_gpu": 1,
@@ -105,7 +105,7 @@ def build_client(model: str, host: str, port: int, *, max_seq_len: int, response
         use_unpad=(os.environ.get("ARCTIC_ATTN_IMPL", "flash_attention_2") == "flash_attention_2"),
         use_autocast=False,
     )
-    cfg = ArcticRLClientConfig(
+    cfg = ArcticClientConfig(
         model_name=model,
         seed=seed,
         max_seq_len=max_seq_len,
@@ -134,7 +134,7 @@ def build_client(model: str, host: str, port: int, *, max_seq_len: int, response
             startup_timeout=startup_timeout,
         ),
     )
-    return SyncArcticRLClient(cfg)
+    return ArcticRLClient(cfg)
 
 
 # --------------------------------------------------------------------------------------------------------------- #
